homematicip_cloud: address joostlek review feedback
- Move async_migrate_entry from migration.py to __init__.py
- Combine obsolete entity removal blocks in migration
- Remove ValueError guard on feature_id (can't happen in production)
- Use entity_registry/device_registry fixtures in tests
- Combine migration tests into single parametrized test
